implement schema keyword
